% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{heat_data}
\alias{heat_data}
\title{Data for L-shape heat transfer problem}
\format{
A list with the following components:
\describe{
\item{exact_qoi}{numeric, QOI at 'exact' solution, i.e.
\code{c(0.1, 1.5, 2.9)}.}
}
}
\usage{
heat_data
}
\description{
Data for L-shape heat transfer problem
}
\keyword{datasets}
