% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_solver.R
\name{mt_const}
\alias{mt_const}
\title{MT problem constants}
\usage{
mt_const(name = c("nparams", "nqoi", "maxiter"))
}
\arguments{
\item{name}{character, constant name}
}
\description{
These are: number of parameters (4), number of quantities of interest (7),
maximum number of solver iterations (8).
}
