% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mt_data}
\alias{mt_data}
\title{Data for MT problem}
\format{
A list with the following components:
\describe{
\item{precision_levels}{available precision levels sorted from the most
inaccurate; each next level increases significantly the time of
computations.}
\item{exact_qoi}{complex, QOI at 'exact' solution, i.e.
\code{c(1, 2, 10, 3)}, computed with precision \code{1.2}.}
}
}
\usage{
mt_data
}
\description{
Data for MT problem
}
\keyword{datasets}
