% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_solver.R
\name{heat_solver}
\alias{heat_solver}
\title{Gateway for 'L-shape' heat transfer problem}
\usage{
heat_solver(wd = getwd(), ...)
}
\arguments{
\item{wd}{character, working directory.}

\item{...}{additional parameters passed to \code{\link[solvergater:shell_solver]{solvergater::shell_solver()}}.}
}
\description{
Problem is 3-dimensional, currently not adaptive.
}
