% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_solver.R
\name{mt_solver}
\alias{mt_solver}
\title{Gateway for Julen Alvarez-Aramberri's magnetotelluric (MT) problem solver}
\usage{
mt_solver()
}
\description{
Beware that the problem is 4-dimensional (x in \code{\link[solvergater:compute_objective]{solvergater::compute_objective()}}
must be numeric vector of length 4). \code{R} code must be run in a directory
containing \code{EXEC/MT} and \verb{FILES/*}.
}
\examples{
\dontrun{
library(solvergater.solvers)
mts <- mt_solver()
compute_objective(mts, c(1, 2, 3, 10), 70)
}
}
