% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_solver.R
\name{mt_solver}
\alias{mt_solver}
\title{Gateway for Julen Alvarez-Aramberri's magnetotelluric (MT) problem solver}
\usage{
mt_solver(wd = getwd(), ...)
}
\arguments{
\item{wd}{character, working directory: the one that contains \code{EXEC/MT}
and \verb{FILES/*}.}

\item{...}{additional params passed to
\code{\link[solvergater:adaptive_shell_solver]{solvergater::adaptive_shell_solver()}}.}
}
\description{
Beware that the problem is 4-dimensional (x in \code{\link[solvergater:run]{solvergater::run()}}
must be numeric vector of length 4).
}
\examples{
\dontrun{
library(solvergater.solvers)
mts <- mt_solver()
nparams(mts)
run(mts, c(1, 2, 3, 10), 70)
}
}
